% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalfunctions.R
\name{acc_changes}
\alias{acc_changes}
\title{Accumulate changes in a LULC time serie raster}
\usage{
acc_changes(path)
}
\arguments{
\item{path}{The path of the raster to be analysed}
}
\value{
Two maps, the first containing the sequence of change as pixel value and second
    the quantity time of change as pixel value.
}
\description{
This fucntion return two maps, the first containing the sequence of change as
pixel value (answering: in what point of time a pixel has changes?) and second
the quantity time of change as pixel value(answering: how many times a pixel
has change). These maps can be plotted  with the plot function method of the
raster package or with the `tmap` package.
}
\examples{
acc_changes(demo_landscape(2000:2005, res = 1, prob = c(0.05, 0.3, 0.05, 0.4, 0.2)))

}
