% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_method.R
\name{$,Interval-method}
\alias{$,Interval-method}
\alias{$,Category-method}
\alias{$,Transition-method}
\title{Accessor method for objects from Intensity Analysis}
\usage{
\S4method{$}{Interval}(x, name)

\S4method{$}{Category}(x, name)

\S4method{$}{Transition}(x, name)
}
\arguments{
\item{x}{the object}

\item{name}{the name of the object}

\item{Interval}{the class}

\item{Category}{the class}

\item{Transition}{the class}
}
\description{
Accessor method for objects from Intensity Analysis
}
\keyword{internal}
