% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherplots.R
\name{alluvial_Land}
\alias{alluvial_Land}
\title{Alluvil diagram of LUC transitions (one or multistep)}
\usage{
alluvial_Land(
  dataset,
  legendtable,
  title = NULL,
  caption = "LUC category",
  xlab = "Year",
  ylab = "Area (km2 or pixel)",
  year_filter = NULL,
  area_km2 = TRUE,
  ...
)
}
\arguments{
\item{dataset}{\code{<tibble>}. A table of the multi step (\code{lulc_Mulstistep})
or one step transitions (\code{lulc_OneStep}) generated by \code{\link{contingencyTable}}.}

\item{legendtable}{\code{<tibble>}. A table (\code{lulc_Onestep}) containing the LUC legend items and their respective color.}

\item{title}{\code{<chr>}. The title of the plot.}

\item{caption}{\code{<chr>}. The caption of the plot.}

\item{xlab}{\code{<chr>}. Label for the x axis.}

\item{ylab}{\code{<chr>}. Label for the y axis.}

\item{year_filter}{\code{<chr>}. A vector containing the years to be ploted.}

\item{area_km2}{\code{<logical>}. If TRUE the change is computed in km2, if FALSE in pixel counts.}

\item{\dots}{additional themes parameters \code{\link[ggplot2]{theme}}}
}
\value{
An alluvial diagram
}
\description{
Alluvial diagram showing the one or multi step LUC transitions during the analysed period.
The difference with sankey is ...
}
