% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherplots.R
\name{barplotLand}
\alias{barplotLand}
\title{A barplot for the area of every point in time of the analysed period}
\usage{
barplotLand(
  dataset,
  legendtable,
  title = "General Area in the interval (2002 - 2014)",
  color = c(GC = "gray70", NG = "#006400", NL = "#EE2C2C"),
  fill = "LULC Classes",
  xlab = "Year",
  ylab = expression(paste("Area ", Km^{     2 })),
  area_km2 = TRUE,
  ...
)
}
\arguments{
\item{dataset}{A table of the multi step transition}

\item{legendtable}{A table contains the legend of the land use classes and his respective color}

\item{title}{The title of the plot}

\item{color}{A color list for the three types of chages}

\item{fill}{Legend}

\item{xlab}{character Label for the x axe}

\item{ylab}{character Label for the y axe}

\item{area_km2}{boolean TRUE for km2 unit, FALSE for pixel unit}

\item{\dots}{themes parameters \code{\link[ggplot2]{theme}}}
}
\value{
a barplot
}
\description{
A barplot for the area of every point in time of the analysed period
}
\seealso{
\code{ggplot2::\link[ggplot2]{theme}}
}
