% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherplots.R
\name{chordDiagramLand}
\alias{chordDiagramLand}
\title{A circlize plot for the transition between two times point, one step.}
\usage{
chordDiagramLand(
  dataset,
  legendtable,
  legposition = c(x = 1.8, y = 1.5),
  legtitle = "Classes",
  sectorcol = "gray80",
  area_km2 = TRUE
)
}
\arguments{
\item{dataset}{A table of the one step transition}

\item{legendtable}{A table contains the legend of the land use classes and his respective color}

\item{legposition}{A list of `x` and `y` parameter position for the legend}

\item{legtitle}{The title of the legen}

\item{sectorcol}{The color of the extern sector containing the years}

\item{area_km2}{boolean TRUE for km2 unit, FALSE for pixel unit}
}
\value{
A Chord Diagram
}
\description{
A circlize plot for the transition between two times point, one step.
}
