% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingencyTable.R
\name{demo_landscape}
\alias{demo_landscape}
\title{Title}
\usage{
demo_landscape(year, nrows = 100, ncols = 100, res = 1, xmn = 0,
  xmx = 100, ymn = 0, ymx = 100, crs = NA, class = 1:5,
  prob = NULL)
}
\arguments{
\item{year}{A list of year, first and last}

\item{nrows}{nrows of the raster}

\item{ncols}{ncols of the raster}

\item{res}{the resolution of the raster to be created}

\item{xmn}{x minimun extent}

\item{xmx}{x maximum extent}

\item{ymn}{y minimum extent}

\item{ymx}{y maximum extent}

\item{crs}{the coordinate referencing system}

\item{class}{the raster classes}

\item{prob}{the probability of occorence for the class list}
}
\description{
Title
}
\examples{
demo_landscape(2000:2006)
}
