% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters_input.R
\name{dot-input_rasters}
\alias{dot-input_rasters}
\alias{.input_rasters,character-method}
\alias{character}
\alias{.input_rasters,list-method}
\alias{list}
\alias{.input_rasters,RasterLayer-method}
\alias{RasterLayer}
\alias{.input_rasters,RasterBrick-method}
\alias{RasterBrick}
\alias{.input_rasters,RasterStack-method}
\alias{RasterStack}
\title{input_rasters}
\usage{
\S4method{.input_rasters}{character}(x, ...)

\S4method{.input_rasters}{list}(x, ...)

\S4method{.input_rasters}{RasterLayer}(x, ...)

\S4method{.input_rasters}{RasterBrick}(x, ...)

\S4method{.input_rasters}{RasterStack}(x, ...)
}
\arguments{
\item{x}{path (character), Raster* object or list of Raster* objects.}

\item{\dots}{additional arguments to \code{raster::\link[raster]{raster}}.}
}
\value{
A RasterStack
}
\description{
A methods for loading the raster into OpenLand
}
\seealso{
\code{raster::\link[raster]{raster}},
\code{raster::\link[raster]{stack}}
}
\keyword{internal}
