% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityClass.R
\docType{class}
\name{Intensity-class}
\alias{Intensity-class}
\title{Class Intensity}
\description{
An S4 super class for the objects from the intensity analysis
}
\section{Slots}{

\describe{
\item{\code{tabela}}{tbl_df.}
}}

