% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityClass.R
\name{intensity}
\alias{intensity}
\title{This is a helper funtion for create the intensities sub classes}
\usage{
intensity(data, color)
}
\arguments{
\item{data}{tibble}

\item{color}{list of color}
}
\value{
an object with the correspondate class
}
\description{
This is a helper funtion for create the intensities sub classes
}
