% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensityanalysis}
\alias{intensityanalysis}
\title{Perform the intensity analysis from a cross-tabulation matrices}
\usage{
intensityanalysis(dataset, class_n, class_m, area_km2 = TRUE)
}
\arguments{
\item{dataset}{The list of object from \code{\link{contingenceTable}}}

\item{class_n}{character. A natural class of land use}

\item{class_m}{character. A antropic class of land use}

\item{area_km2}{logical. Whether the change would be computed in km2 or quantity of pixel unit}
}
\value{
Intensity objects including the levels tables and strationarity test tables
}
\description{
This function inplement the Intensity Analysis, a quantitative method to analyze
maps of land categories from several points in time for a single site by
considering cross-tabulation matrices, where one matrix summarizes the change
in each time interval.
}
\details{
There are three levels of analysis, starting from general to more detailed levels, where each level
exposes different types of information given the previous level of analysis
\cite{(Aldwaik and Pontius, 2012, 2013)}.

\enumerate{
 \item The \emph{interval level} examines how the size and speed of change vary
across time intervals.
 \item The \emph{category level} examines how the size and intensity of gross losses
and gross gains in each category vary across categories for each time interval.
 \item The \emph{transition level} examines how the size and intensity of a category’s
transitions vary across the other categories that are available for that transition.
  }

At each level, the method tests for stationarity of patterns across time intervals.

\bold{The function return a list with 10 objects:}
\itemize{
 \item lulc_table: A \code{tibble} of contingency between all step of time analysed, contains 8 columns:
   \enumerate{
   \item Period:  \code{<fct>}. The period of change in the format \code{first year - last year}.
   \item From: \code{<fct>}. The class in the first year.
   \item To: \code{<fct>}. The class in the last year.
   \item km2: \code{dbl}. The quantity in kilometer that transit from the classes \code{From}
   to \code{To} in the period.
   \item interval: \code{dbl}. The number of time point between the first and the last year of the period.
   \item QtPixel: \code{int}. The quantity in number of pixel that transits from
   the classes \code{From} to \code{To} in the period.
   }
 \item lv1_tbl: An \code{"\linkS4class{IntensityL01}"} object containing
 1 table with 5 columns:
   \enumerate{
   \item Period: \code{<fct>}.
   \item intch_km2: \code{<dbl>}. Change in kilometer or quantity of pixel in the period.
   \item St: \code{<dbl>}. annual intensity of change for time interval [Yt, Yt+1].
   \item U: \code{<dbl>}. value of uniform line for time intensity analysis.
   \item tipo: \code{<chr>}. The type of change Fast when St > U and Slow when St < U.
   }
 \item lv2_gain: An \code{"\linkS4class{IntensityL02}"} object containing
 a slot with lookup table af the colors associate with the land classes and a
 table with 8 columns:
 \enumerate{
   \item Period: \code{<fct>}.
   \item To: \code{<fct>}. a category j.
   \item interval: \code{<dbl>}. duration of [Yt, Yt+1].
   \item num03: \code{<dbl>}. area of gross gain of category j during [Yt, Yt+1]/duration of [Yt, Yt+1].
   \item denom03: \code{<dbl>}. area of category j at time Yt+1.
   \item Gtj: \code{<dbl>}. annual intensity of gross gain of category j for time interval [Yt, Yt+1].
   \item intch_km2 or QtPixel: \code{<dbl> or <int>}.
   \item St: \code{<dbl>}. annual intensity of change for time interval [Yt, Yt+1].
   }
 \item lv2_loss: An \code{"\linkS4class{IntensityL02}"} object containing
 a slot with lookup table af the colors associate with the land classes and a
 table with 8 columns:
 \enumerate{
   \item Period: \code{<fct>}.
   \item From: \code{<fct>}.
   \item interval: \code{<dbl>}.
   \item num04: \code{<dbl>}.
   \item denom04: \code{<dbl>}.
   \item Lti: \code{<dbl>}.
   \item intch_km2 or QtPixel: \code{<dbl> or <int>}.
   \item STt: \code{<dbl>}.
   }
 \item gain_n: An \code{"\linkS4class{IntensityL03}"} object containing
 a slot with lookup table af the colors associate with the land classes and a
 table with 10 columns:
 \enumerate{
   \item Period: \code{<fct>}.
   \item From: \code{<fct>}.
   \item interval: \code{<dbl>}.
   \item num05: \code{<dbl>}.
   \item denom05: \code{<dbl>}.
   \item Rtin: \code{<dbl>}.
   \item To: \code{<fct>}.
   \item num06: \code{<dbl>}.
   \item denom06: \code{<dbl>}.
   \item Wtn: \code{<dbl>}.
   }
 \item loss_m: An \code{"\linkS4class{IntensityL03}"} object containing
 a slot with lookup table af the colors associate with the land classes and a
 table with 10 columns:
 \enumerate{
   \item Period: \code{<fct>}.
   \item To: \code{<fct>}.
   \item interval: \code{<dbl>}.
   \item num07: \code{<dbl>}.
   \item denom07: \code{<dbl>}.
   \item Qtmj: \code{<dbl>}.
   \item From: \code{<fct>}.
   \item num08: \code{<dbl>}.
   \item denom08: \code{<dbl>}.
   \item Vtn: \code{<dbl>}.
   }
 \item st_lv2_gain: A \code{tibble} of stationarity test in level 02 gain
 containing 5 columns:
 \enumerate{
   \item To: \code{<fct>}.
   \item gain: \code{<int>}.
   \item N: \code{<int>}.
   \item Stationarity: \code{<chr>}.
   \item Test: \code{<chr>}.
   }
 \item st_lv2_loss: A \code{tibble} of stationarity test in level 02 loss
 containing 5 columns:
 \enumerate{
   \item From: \code{<fct>}.
   \item loss: \code{<int>}.
   \item N: \code{<int>}.
   \item Stationarity: \code{<chr>}.
   \item Test: \code{<chr>}.
   }
 \item st_gain_n: A \code{tibble} of stationarity test in level 03 gain on class n
 containing 5 columns:
 \enumerate{
   \item From: \code{<fct>}.
   \item loss: \code{<int>}.
   \item N: \code{<int>}.
   \item Stationarity: \code{<chr>}.
   \item Test: \code{<chr>}.
   }
 \item st_loss_m: A \code{tibble} of stationarity test in level 03 loss on class m
 containing 5 columns:
 \enumerate{
   \item To: \code{<fct>}. The classes gaining from the m class.
   \item gain: \code{<int>}. Number of time the class gain from the m class.
   \item N: \code{(<int>)}. Number total of time of trasition to be considered as stationary.
   \item Stationarity: \code{<chr>}. \emph{targeted} or \emph{avoided} by the \code{m} class.
   \item Test: \code{<chr>}. \emph{Y} for stationarity detected and \emph{N} when not.
   }
  }
}
\examples{
intensityanalysis(dataset = SL_2002_2014, class_n = "Ap", class_m = "SG", area_km2 = TRUE)


}
\references{
Aldwaik, S. Z. and Pontius, R. G. (2012) ‘Intensity analysis to unify
measurements of size and stationarity of land changes by interval, category, and
transition’, Landscape and Urban Planning. Elsevier B.V., 106(1), pp. 103–114.
\doi{10.1016/j.landurbplan.2012.02.010}.

Aldwaik, S. Z. and Pontius, R. G. (2013) ‘Map errors that could account for deviations
from a uniform intensity of land change’, International Journal of Geographical
Information Science. Taylor & Francis, 27(9), pp. 1717–1739. \doi{10.1080/13658816.2013.787618}.
}
