% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityClass.R
\docType{class}
\name{IntensityL01-class}
\alias{IntensityL01-class}
\title{Class Intensity Level 1}
\description{
An S4 class for the intensity level 1 analysis result
}
\section{Slots}{

\describe{
\item{\code{level}}{character}

\item{\code{intervalData}}{a tibble}
}}

