% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\docType{methods}
\name{plot01}
\alias{plot01}
\alias{plot,IntensityL01-method}
\title{A method in the generic plot for the openland class IntensityL01}
\usage{
\S4method{plot}{IntensityL01,missing}(x, y, labels = c(leftlabel =
  bquote("Mudancas de area" ~ "(" ~ Km^2 ~ ")"), rightlabel =
  "Taxa de mudança (\%)", title = "Level 01 Mudanças todo o Periodo"),
  labs = c(type = "Mudanças", ur = "U"), marginplot = c(lh = -10, rh =
  0), leg_curv = c(x = 1/10, y = 1/10), color_bar = c(faster =
  "#B22222", slower = "#006400", area = "gray40"), ...)
}
\arguments{
\item{x}{IntensityL01.}

\item{y}{missing.}

\item{labels}{character}

\item{labs}{character}

\item{marginplot}{character}

\item{leg_curv}{character}

\item{color_bar}{character}

\item{\dots}{additional arguments}

\item{leftlabel}{character}

\item{rightlabel}{character}

\item{title}{character}

\item{type}{character}

\item{ur}{character}
}
\value{
intensity graph
}
\description{
A method in the generic plot for the openland class IntensityL01
}
