% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityClass.R
\docType{class}
\name{IntensityL01}
\alias{IntensityL01}
\title{An S4 class for the intensity level 1 analysis}
\value{
an intensity level 1 object
}
\description{
An S4 class for the intensity level 1 analysis
}
\section{Slots}{

\describe{
\item{\code{level}}{character.}
}}

