% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityClass.R
\docType{class}
\name{IntensityL02-class}
\alias{IntensityL02-class}
\title{Class Intensity Level 2}
\description{
A class for the intensity level 2 analysis result
}
\section{Slots}{

\describe{
\item{\code{level}}{character}

\item{\code{lookupcolor}}{character}

\item{\code{categoryData}}{a tibble}

\item{\code{categoryStationarity}}{a tibble}
}}

