% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\docType{methods}
\name{plot02}
\alias{plot02}
\alias{plot,IntensityL02-method}
\title{Title: setting a method in the generic plot for the openland class IntensityL02}
\usage{
\S4method{plot}{IntensityL02,missing}(x, y, labels = c(leftlabel =
  bquote("Ganho de area (" ~ Km^2 ~ ")"), rightlabel =
  "Intensidade de ganho (\%)", title = "Level 02 Ganho Anual"),
  labs = c(type = "Classes", ur = "STt"), marginplot = c(lh = 0.5, rh =
  0.5), leg_curv = c(x = 1/10, y = 1/10), ...)
}
\arguments{
\item{x}{IntensityL02.}

\item{y}{missing.}

\item{labels}{character}

\item{labs}{character}

\item{marginplot}{character}

\item{leg_curv}{character}

\item{\dots}{additional arguments}

\item{rightlabel}{character}

\item{leftlabel}{character}

\item{title}{character}

\item{type}{character}

\item{ur}{character}

\item{lh}{character}

\item{rh}{character}
}
\value{
intensity graph
}
\description{
Title: setting a method in the generic plot for the openland class IntensityL02
}
