% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityClass.R
\docType{class}
\name{IntensityL02}
\alias{IntensityL02}
\title{A class for the intensity level 2 analysis}
\value{
an intensity level 2 object
}
\description{
A class for the intensity level 2 analysis
}
\section{Slots}{

\describe{
\item{\code{level}}{character.}

\item{\code{color}}{character.}
}}

