% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityClass.R
\docType{class}
\name{IntensityL03-class}
\alias{IntensityL03-class}
\title{Class Intensity Level 3}
\description{
An S4 class for the intensity level 2 analysis
}
\section{Slots}{

\describe{
\item{\code{level}}{character}

\item{\code{color}}{character}
}}

