% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityClass.R
\docType{class}
\name{Interval-class}
\alias{Interval-class}
\alias{Interval}
\title{Class Interval}
\description{
An S4 class for the Interval level analysis result
}
\details{
\bold{The slot \code{intervalData} receives a table containing 4 columns in the following format:}
   \enumerate{
   \item Period: \code{<fct>}. The period \emph{[Yt, Yt+1]}.
   \item PercentChange: \code{<dbl>}. Interval Change Area (percent of map).
   \item St: \code{<dbl>}. annual intensity of change for time
   interval [Yt, Yt+1].
   \item U: \code{<dbl>}. value of uniform line for time intensity analysis.
   }
}
\section{Slots}{

\describe{
\item{\code{intervalData}}{tibble; A table of Interval level's result (\emph{St} and \emph{U} values)}
}}

