% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalfunctions.R
\name{lulc_gain}
\alias{lulc_gain}
\title{Gain and persistence area in a LULC raster time series}
\usage{
lulc_gain(path, persistence = 99)
}
\arguments{
\item{path}{list. List of filenames, list of Raster* objects, RasterBrick(\code{\link[raster]{brick}}) or
RasterStack(\code{\link[raster]{stack}})}

\item{persistence}{int. value to be assigned to a persistence pixel}
}
\value{
A RasterBrick
}
\description{
This function calculates the gain of every category during the analysed period.
For a series containing n rasters it returns a list of (n - 1) rasters, i.e, a
raster for every interval.
}
\examples{
\donttest{gain_map(SaoLourencoBasin)}
}
