% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherplots.R
\name{netgrossplot}
\alias{netgrossplot}
\title{A Bar Plot}
\usage{
netgrossplot(
  dataset,
  legendtable,
  title = "General Changes (First year - Last year)",
  xlab = "Land Use Classes",
  ylab = "Area (Kilometer)",
  changesLabel = c(GC = "Gross changes", NG = "Net Gain", NL = "Net Loss"),
  color = c(GC = "gray70", NG = "#006400", NL = "#EE2C2C"),
  area_km2 = TRUE
)
}
\arguments{
\item{dataset}{\code{<tibble>}. A table of the multi step transition (\code{lulc_Mulstistep}) from \code{\link{contingencyTable}}.}

\item{legendtable}{\code{<tibble>}. A table (\code{lulc_Onestep}) contains the legend of the LUC category and his respective color.}

\item{title}{\code{<character>}. The title of the plot (optional), use \code{NULL} for no title.}

\item{xlab}{\code{<character>}. Label for the x axis}

\item{ylab}{\code{<character>}. Label for the y axis}

\item{changesLabel}{\code{<character>}. Labels for the three types of chages, the
defaults are c(GC = "Gross chages", NG = "Net Gain", NL = "Net Loss")}

\item{color}{\code{<chr>}. A vector of the the three types of changes colors.}

\item{area_km2}{\code{<logical>}. If TRUE the change is computed in km2, if FALSE in pixel counts.}
}
\value{
A bar plot
}
\description{
A barplot showing the LUC evolution trough the years, it presents the Net Gain/Loss
and Gross Changes between land use classes for the period analysed.
}
\examples{
test1 <- demo_landscape(2000:2005)
test2 <- contingencyTable(input_raster = test1, pixelresolution = 1)
netgrossplot(dataset = test2$lulc_Multistep, legendtable = test2$tb_legend, area_km2 = FALSE)


}
