% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\docType{methods}
\name{plot,IntensityL01,missing-method}
\alias{plot,IntensityL01,missing-method}
\title{Title: setting a method in the generic plot for the openland class IntensityL01}
\usage{
\S4method{plot}{IntensityL01,missing}(x, y = "missing",
  labels = c(leftlabel = bquote("Mudancas de area" ~ "(" ~ Km^2 ~ ")"),
  rightlabel = "Taxa de mudança (\%)", title =
  "Level 01 Mudanças todo o Periodo"), labs = c(type = "Mudanças", ur =
  "U"), marginplot = c(lh = -10, rh = 0), leg_curv = c(x = 1/10, y =
  1/10), color_bar = c(faster = "#B22222", slower = "#006400", area =
  "gray40"), ...)
}
\arguments{
\item{x}{IntensityL01.}

\item{y}{missing.}
}
\description{
Title: setting a method in the generic plot for the openland class IntensityL01
}
