% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\docType{methods}
\name{plot,IntensityL02,missing-method}
\alias{plot,IntensityL02,missing-method}
\title{Title: setting a method in the generic plot for the openland class IntensityL02}
\usage{
\S4method{plot}{IntensityL02,missing}(x, y = "missing",
  labels = c(leftlabel = bquote("Ganho de area (" ~ Km^2 ~ ")"),
  rightlabel = "Intensidade de ganho (\%)", title =
  "Level 02 Ganho Anual"), labs = c(type = "Classes", ur = "STt"),
  marginplot = c(lh = 0.5, rh = 0.5), leg_curv = c(x = 1/10, y = 1/10),
  ...)
}
\arguments{
\item{x}{IntensityL02.}

\item{y}{missing.}
}
\description{
Title: setting a method in the generic plot for the openland class IntensityL02
}
