% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\docType{methods}
\name{plot,IntensityL03,missing-method}
\alias{plot,IntensityL03,missing-method}
\title{Title: setting a method in the generic plot for the openland class IntensityL03}
\usage{
\S4method{plot}{IntensityL03,missing}(x, y, labels = c(rightlabel =
  bquote("Intensidade de ganho da classe" ~
  .(as.character(dataset[[7]][[1]])) ~ "(\%)"), leftlabel =
  bquote("Ganho da classe" ~ .(as.character(dataset[[7]][[1]])) ~ "(" ~
  Km^2 ~ ")"), title = "Level 03 Ganho da Ap"), labs = c(type =
  "Classes", ur = "Wtn/Vtm"), marginplot = c(lh = 0.5, rh = 0.5),
  leg_curv = c(x = 1/10, y = 1/10), ...)
}
\arguments{
\item{x}{IntensityL03.}

\item{y}{missing.}
}
\description{
Title: setting a method in the generic plot for the openland class IntensityL03
}
