% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{plot}
\alias{plot}
\alias{plot,Interval,ANY-method}
\alias{plot,Category,ANY-method}
\alias{plot,Transition,ANY-method}
\title{Plot method for objects from Intensity Analysis}
\usage{
\S4method{plot}{Interval,ANY}(
  x,
  y,
  labels = c(leftlabel = "Interval Change Area (percent of map)", rightlabel =
    "Annual Change Area (percent of map)", title = "Level 01 Mudancas todo o Periodo"),
  labs = c(type = "Changes", ur = "Uniform Rate"),
  marginplot = c(lh = -10, rh = 0),
  leg_curv = c(x = 1/10, y = 1/10),
  color_bar = c(faster = "#B22222", slower = "#006400", area = "gray40"),
  ...
)

\S4method{plot}{Category,ANY}(
  x,
  y,
  labels = c(leftlabel = bquote("Gain/Loss area (" ~ Km^2 ~ ")"), rightlabel =
    "Intensity Gain/Loss (\%)", title = "Level 02 Ganho/Perda Anual"),
  labs = c(type = "Classes", ur = "Uniform Rate"),
  marginplot = c(lh = 0.5, rh = 0.5),
  leg_curv = c(x = 1/10, y = 1/10),
  ...
)

\S4method{plot}{Transition,ANY}(
  x,
  y,
  labels = c(rightlabel = bquote("Intensity Gain/Loss of" ~
    .(as.character(dataset[[3]][[1]])) ~ "(\%)"), leftlabel = bquote("Gain/Loss of " ~
    .(as.character(dataset[[3]][[1]])) ~ "(" ~ Km^2 ~ ")"), title =
    "Level 03 Ganho da classe m/n"),
  labs = c(type = "Classes", ur = "Uniform Rate"),
  marginplot = c(lh = 0.5, rh = 0.5),
  leg_curv = c(x = 1/10, y = 1/10),
  ...
)
}
\arguments{
\item{x}{tibble. An object from the intensity analysis data}

\item{y}{not used}

\item{labels}{character. left right and title}

\item{labs}{character}

\item{marginplot}{character}

\item{leg_curv}{character}

\item{color_bar}{character}

\item{\dots}{additional arguments themes parameters \code{\link[ggplot2]{theme}}}

\item{Interval}{the class}

\item{Category}{the class}

\item{Transition}{the class}
}
\value{
An intensity graph
}
\description{
Plot \code{Intensity} objects based on Intensity Analysis tables
}
