% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityClass.R
\docType{methods}
\name{$}
\alias{$}
\alias{$,Intensity-method}
\title{replace parts of Intensity}
\usage{
\S4method{$}{Intensity}(x, name)
}
\arguments{
\item{x}{tibble}

\item{name}{character}
}
\description{
replace parts of Intensity
}
