% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherplots.R
\name{sankeyLand}
\alias{sankeyLand}
\title{Sankey diagram of LUC transitions (one or multistep)}
\usage{
sankeyLand(dataset, legendtable, iterations = 0)
}
\arguments{
\item{dataset}{\code{<tibble>}. A table of the multi step (\code{lulc_Mulstistep})
or one step transitions (\code{lulc_OneStep}) generated by \code{\link{contingencyTable}}.}

\item{legendtable}{\code{<tibble>}. A table (\code{lulc_Onestep}) containing the LUC legend items and their respective color.}

\item{iterations}{numeric. Number of iterations in the diagram layout for computation of the depth (y-position) of each node.
\code{\link[networkD3]{sankeyNetwork}}.}
}
\value{
A sankey diagram
}
\description{
A sankey showing the one or multi step LUC transitions during the analysed period.
}
\seealso{
\code{\link[networkD3]{sankeyNetwork}}
}
