% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SaoLourencoBasin}
\alias{SaoLourencoBasin}
\title{Land use and cover (LUC) rasters of the São Lourenço River Basin (2002 - 2014)}
\format{An object of class \code{"RasterBrick"} with 5 layers:
\itemize{
  \item \strong{landscape_2002:} LUC in 2002
  \item \strong{landscape_2008:} LUC in 2008
  \item \strong{landscape_2010:} LUC in 2010
  \item \strong{landscape_2012:} LUC in 2012
  \item \strong{landscape_2014:} LUC in 2014
  }}
\source{
\url{https://www.embrapa.br/pantanal/bacia-do-alto-paraguai}
}
\usage{
data(SaoLourencoBasin)
}
\description{
A multi-layer raster (RasterStack) with five (5) RasterLayer objects
representing LUC in the years of 2002, 2008, 2010, 2012 and 2014 respectively.
Each Raster* has 11 LUC classes:
\tabular{cclll}{
\strong{Pixel Value} \tab \strong{Legend} \tab \strong{Class}
\tab \strong{Use} \tab \strong{Details}  \cr
2        \tab Ap     \tab Anthropic \tab Anthropic Use \tab Pastagem    \cr
3        \tab FF     \tab Natural   \tab NA            \tab Formações Florestais \cr
4        \tab SA     \tab Natural   \tab NA            \tab Savana Arborizada (SA - Cerrado)  \cr
5        \tab SG     \tab Natural   \tab NA            \tab Savana Gramínea (SG - Campo) \cr
7        \tab aa     \tab Anthropic \tab NA            \tab Alteração Antrópica \cr
8        \tab SF     \tab Natural   \tab NA            \tab Savana Florestada (SF - Cerradão) \cr
9        \tab Agua   \tab Natural   \tab NA            \tab Rios, córregos, corixos, vazantes, baías e salinas   \cr
10       \tab Iu     \tab Anthropic \tab Anthropic Use \tab Influencia Urbana   \cr
11       \tab Ac     \tab Anthropic \tab Anthropic Use \tab Agricultura           \cr
12       \tab R      \tab Anthropic \tab Anthropic Use \tab Reflorestamento   \cr
13       \tab Im     \tab Anthropic \tab Anthropic Use \tab Degradação por Mineração
}
}
\examples{
data(SaoLourencoBasin)
\donttest{raster::plot(SaoLourencoBasin[[1]])}

}
\references{
Instituto SOS Pantanal and WWF-Brasil (2015)
(\href{https://d3nehc6yl9qzo4.cloudfront.net/downloads/publicacao_bap_relatorio_2012_2014_web.pdf}{Monitoramento
das alterações da cobertura vegetal e uso do Solo na Bacia do
Alto Paraguai – Porção Brasileira – Período de Análise: 2012 a 2014. 4th edn. Brasilia.})
}
\keyword{datasets}
