% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalfunctions.R
\name{summary_dir}
\alias{summary_dir}
\title{Summary of multiple parameters in a raster directory}
\usage{
summary_dir(path)
}
\arguments{
\item{path}{The path for a raster directory/list to be analysed}
}
\value{
Table with the raster parameters in columns
}
\description{
Listing major charateristics of raster inputs
}
\examples{

summary_dir(demo_landscape(2000:2005, res = 1, prob = c(0.05, 0.3, 0.05, 0.4, 0.2)))


}
