% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityClass.R
\docType{class}
\name{Transition-class}
\alias{Transition-class}
\alias{Transition}
\title{Class Transition}
\description{
An S4 class for the Transition level analysis  result, the objects of this class
can be plot with the plot method \code{\link{plot}}
}
\details{
\bold{The slots \code{transitionData} and \code{transitionStationarity} can receive
   tables for "Gain of class n" or "Loss of class m" in the following format:}
\enumerate{

\item Gain of class n:

 \itemize{

  \item \code{transitionData}: \code{<tibble>}. A table with 7 columns:
  \enumerate{
   \item Period: \code{<fct>}. The period \emph{[Yt, Yt+1]}.
   \item From: \code{<fct>}. a category i.
   \item To: \code{<fct>}. the gaining category in the transition of interest \emph{(n)}.
   \item Interval: \code{<int>}. duration of the period \emph{[Yt, Yt+1]}.
   \item T_i2n_km2/T_i2n_pixel: \code{<dbl>}. transition from category i to category n
   during time interval \emph{[Yt, Yt+1]} where \emph{i}  \code{is not equal to} \emph{n}.
   \item Rtin: \code{<dbl>}. annual intensity of transition from category i to category n
   during time interval \emph{[Yt, Yt+1]} where \emph{i}  \code{is not equal to} \emph{n}.
   \item Wtn: \code{<dbl>}. value of uniform intensity of transition to category
   n from all non-n categories at time Yt during time interval \emph{[Yt, Yt+1]}.
   }
 \item transitionStationarity: \code{<tibble>}. A table of stationarity test over the gain on \emph{class n}
 containing 5 columns:
 \enumerate{
   \item From: \code{<fct>}. the losing category in the transition of interest to the n category.
   \item loss: \code{<int>}. Number of time the class loss to the n class.
   \item N: \code{<int>}. Number total of time of trasition to be considered as stationary (T).
   \item Stationarity: \code{<chr>}. \emph{targeted by} or \emph{avoided by} by the \code{n} class.
   \item Test: \code{<chr>}. \emph{Y} for stationarity detected and \emph{N} when not.
   }
   }

\item Loss of class m:

 \itemize{

  \item \code{transitionData}: \code{<tibble>}. A table with 7 columns:
  \enumerate{
   \item Period: \code{<fct>}. The period \emph{[Yt, Yt+1]}.
   \item To: \code{<fct>}. a category \emph{j}.
   \item From: \code{<fct>}. the losing category in the transition of interest (m).
   \item Interval: \code{<dbl>}. duration of the period \emph{[Yt, Yt+1]}.
   \item T_m2j_km2/T_m2j_pixel: \code{<dbl>}. a transition from category \emph{m} to
   category \emph{j} during time interval \emph{[Yt, Yt+1]}
   where \emph{j}  \code{is not equal to} \emph{m}.
   \item Qtmj: \code{<dbl>}. annual intensity of transition from category \emph{m} to
   category \emph{j} during time interval \emph{[Yt, Yt+1]}
   where \emph{j}  \code{is not equal to} \emph{m}.
   \item Vtm: \code{<dbl>}. value of uniform intensity of transition from
   category \emph{m} to all \emph{non-m} categories at time \out{Y<sub>t+1</sub>}
   during time interval \emph{[Yt, Yt+1]}.
   }
 \item transitionStationarity: \code{<tibble>}. A table of stationarity test over the loss of \emph{class m}
 containing 5 columns:
 \enumerate{
   \item To: \code{<fct>}. the gaining category in the transition of interest from the m category.
   \item gain: \code{<int>}. Number of time the class gain from the m class.
   \item N: \code{<int>}. Number total of time of trasition to be considered as stationary (T).
   \item Stationarity: \code{<chr>}. \emph{targeted} or \emph{avoided} by the \code{m} class.
   \item Test: \code{<chr>}. \emph{Y} for stationarity detected and \emph{N} when not.
   }
   }

}
}
\section{Slots}{

\describe{
\item{\code{lookupcolor}}{character; The colors associate with the data with the legends as name’s attribute}

\item{\code{transitionData}}{tibble; A table of Transition level's result (gain n (\emph{Rtin} & \emph{Wtn})
or loss m (\emph{Qtmj} & \emph{Vtm}) values)}

\item{\code{transitionStationarity}}{tibble; A table containing a stationarity test. A change
is take as stationary only whether the intensities for all time intervals reside
on one side of the uniform intensity, i.e, smaller or bigger than the uniform
rate over the whole period.}
}}

