% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalfunctions.R
\name{acc_changes}
\alias{acc_changes}
\title{Accumulates changes in a LULC raster time series}
\usage{
acc_changes(path)
}
\arguments{
\item{path}{list. List of filenames, list of Raster* objects, RasterStack(\code{\link[raster]{brick}}) or
RasterStack(\code{\link[raster]{stack}})}
}
\value{
Two objects, a raster layer and a table.
}
\description{
This function calculates the number of times a pixel has changed during the
analysed period. It returns a raster with the number of changes as pixel value
and a table containing the areal percentage of every pixel value (number of chages).
}
\examples{
test <- demo_landscape(2000:2005, res = 1, prob = c(0.05, 0.3, 0.05, 0.4, 0.2))
acc_changes(test)
acc_changes(raster::stack(test))
acc_changes(raster::brick(test))

}
