% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherplots.R
\name{anualplot}
\alias{anualplot}
\title{A barplot for the area of every point in time of the analysed period}
\usage{
anualplot(dataset, legendtable,
  title = "General Area in the interval (2002 - 2014)", color = c(GC =
  "gray70", NG = "#006400", NL = "#EE2C2C"))
}
\arguments{
\item{dataset}{A table of the multi step transition}

\item{legendtable}{A table contains the legend of the land use classes and his respective color}

\item{title}{The title of the plot}

\item{color}{A color list for the three types of chages}
}
\value{
a barplot
}
\description{
A barplot for the area of every point in time of the analysed period
}
