% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherplots.R
\name{barplotLand}
\alias{barplotLand}
\title{Area of LUC categories at time points}
\usage{
barplotLand(
  dataset,
  legendtable,
  title = "LUC area in the interval (2002 - 2014)",
  caption = "LUC category",
  xlab = "Year",
  ylab = expression(paste("Area ", km^{     2 })),
  area_km2 = TRUE,
  ...
)
}
\arguments{
\item{dataset}{\code{<tibble>}. A table of the multi step transitions (\code{lulc_Mulstistep}) generated by \code{\link{contingencyTable}}.}

\item{legendtable}{\code{<tibble>}. A table (\code{lulc_Onestep}) containing the LUC legend items and their respective color.}

\item{title}{\code{<chr>}. The title of the plot}

\item{caption}{\code{<chr>}. The caption of the plot.}

\item{xlab}{\code{<chr>}. Label for the x axis.}

\item{ylab}{\code{<chr>}. Label for the y axis.}

\item{area_km2}{\code{<logical>}. If TRUE the change is computed in km2, if FALSE in pixel counts.}

\item{\dots}{additional themes parameters \code{\link[ggplot2]{theme}}}
}
\value{
a barplot
}
\description{
A grouped barplot representing the areas of LUC categories at each time point of the analysed period
}
\seealso{
\code{ggplot2::\link[ggplot2]{theme}}
}
