% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherplots.R
\name{chordDiagramLand}
\alias{chordDiagramLand}
\title{A circlize plot for the transition between two times point, one step.}
\usage{
chordDiagramLand(
  dataset,
  legendtable,
  legposition = c(x = 1.8, y = 1.5),
  legtitle = "Classes",
  sectorcol = "gray80",
  area_km2 = TRUE
)
}
\arguments{
\item{dataset}{\code{<tibble>}. A table of the one step transition (\code{lulc_OneStep}) from \code{\link{contingencyTable}}.}

\item{legendtable}{\code{<tibble>}. A table (\code{lulc_Onestep}) contains the legend of the LUC category and his respective color.}

\item{legposition}{\code{<numeric>}. A vector containing the `x` and `y` values for the position of the legend.}

\item{legtitle}{\code{<character>}. The title of the legend.}

\item{sectorcol}{\code{<character>}. The color of the extern sector containing the years.}

\item{area_km2}{\code{<logical>}. If TRUE the change is computed in km2, if FALSE in pixel counts.}
}
\value{
A Chord Diagram
}
\description{
A circlize plot for the transition between two times point, one step.
}
