% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingencyTable.R
\name{contingenceTable}
\alias{contingenceTable}
\title{Contingence Table}
\usage{
contingenceTable(input_raster, pixelresolution = 30)
}
\arguments{
\item{input_raster}{a raster list}

\item{pixelresolution}{the pixel resolution in meter}
}
\value{
a list of table contengency
}
\description{
Contingence Table
}
\examples{

contingenceTable(demo_landscape(2000:2006), 1)
}
