% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingencyTable.R
\name{contingenceTable}
\alias{contingenceTable}
\title{Contingence Table}
\usage{
contingenceTable(input_raster, pixelresolution = 30)
}
\arguments{
\item{input_raster}{list (of filenames or Raster* objects),
RasterStack(\code{\link[raster]{brick}}) or RasterStack(\code{\link[raster]{stack}})}

\item{pixelresolution}{numeric. The pixel resolution in meter.}
}
\value{
A list that contains 5 objects.
\itemize{
  \item \code{lulc_Mulstistep}: \code{<tibble>}  The table of contingency between all step of time analysed, contains 8 columns:
  \enumerate{
  \item Period: \code{<chr>} The period \emph{[Yt, Yt+1]}.
  \item From: \code{<dbl>} a category \emph{i}.
  \item To: \code{<dbl>} a category \emph{j}.
  \item km2: \code{<dbl>} The quantity in kilometer that transit from the classes category \emph{i}
   to category \emph{j} in the period \emph{[Yt, Yt+1]}.
  \item Interval: \code{<dbl>} The number of time point between the first and
   the last year of the period \emph{[Yt, Yt+1]}.
  \item QtPixel: \code{<int>} The quantity of pixel that transit from the classes category \emph{i}
   to category \emph{j} in the period \emph{[Yt, Yt+1]}.
  \item yearFrom: \code{<chr>} The year that the change come from \emph{[Yt]}
  \item yearTo: \code{<chr>} The year that the change go for \emph{[Yt+1]}
  }
  \item \code{lulc_Onestep}:\code{<tibble>} The table of contingency between the
  first and the last year analysed \emph{[Yt, Yt+1]}, contains
  8 columns like \code{lulc_Mulstistep}.
  \item \code{tb_legend}: \code{<tibble>} A table of the pixel value his name and color, contains 3 columns:
  \enumerate{
  \item classValue: \code{<dbl>} the pixel value of the classes of land use
  \item className: \code{<factor>} The name or legend associate with a given pixel value, here as factor with a specific level for blablabla
  \item color: \code{<chr>} The color associate with the given pixel value.
  Befor others step of analysis, one would want to change the \code{className} and \code{color} values. Some details.
    \itemize{
        \item the class name have to be  in the same order of the \code{classValue}
        column, the \code{levels} the order that the graphs legend will have to be printed. Like:
        \preformatted{

        myobject$tb_legend$className <- factor(c("name1", "name2", "name3", "name4"),
                                               levels = c("name3", "name2", "name1", "name4"))}
        \item the color by the same order of the \code{classValue} comuln, it can be a
       colour name (eg. "black") or an HEX value (eg. #FFFFFF). Like:
       \preformatted{

       myobject$tb_legend$color <- c("#CDB79E", "red", "#66CD00", "yellow")}}}
  \item \code{totalArea}: \code{<tibble>}  A table of two columns with the total area of study, contains 2 columns:
  \enumerate{
  \item area_km2: \code{<numeric>} The total area of study in square meter.
  \item QtPixel: \code{<numeric>} The total area of study in quantity of pixel
  }
  \item \code{totalInterval}: \code{<numeric>} The value represinting the whole time points
  interval of time analysed \emph{[Yt, Yt+1]}, in years.
  }
}
\description{
Contingence Table
}
\examples{

contingenceTable(demo_landscape(year = 2000:2005, res = 1), pixelresolution = 1)
}
