% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demolandscape.R
\name{demo_landscape}
\alias{demo_landscape}
\title{Create Raster with Random pixel Value}
\usage{
demo_landscape(
  year,
  nrows = 100,
  ncols = 100,
  res = 1,
  xmn = 0,
  xmx = 100,
  ymn = 0,
  ymx = 100,
  crs = NA,
  class = 1:5,
  prob = NULL,
  sim.autocorrelate = FALSE,
  bin.interval = TRUE,
  psill = 0.025,
  model = "Exp",
  range = 1000,
  variance = c("identity", "mu", "mu(1-mu)")[1]
)
}
\arguments{
\item{year}{A list of year, first and last included}

\item{nrows}{nrows of the raster to be created}

\item{ncols}{ncols of the raster to be created}

\item{res}{the resolution of the raster to be created}

\item{xmn}{x minimun extent}

\item{xmx}{x maximum extent}

\item{ymn}{y minimum extent}

\item{ymx}{y maximum extent}

\item{crs}{the coordinate referencing system}

\item{class}{the raster classes}

\item{prob}{the probability of occorence for the class list}

\item{sim.autocorrelate}{logical This argument define if the simulation will
be a autocorrelate in space or pixel of independent values.}

\item{bin.interval}{logical This argument define if the categorisation of the
value will be via igual interval or igual number see `ggplot2`}

\item{psill}{sill see variogram}

\item{model}{a model see variogram}

\item{range}{a range of the variability}

\item{variance}{character; variance function to transform to non-stationary
covariances; "identity" does not transform, other options are "mu"
(Poisson) and "mu(1-mu)" (binomial)}
}
\value{
list
}
\description{
This function create a raster serie with some setup like the layer name and
the sample value for the lulc
}
\examples{
demo_landscape(2000:2005, res = 1)
}
\keyword{internal}
