% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensityanalysis}
\alias{intensityanalysis}
\title{It's a function that perform the intensity test analysis from a table contingency}
\usage{
intensityanalysis(dataset, class_n, class_m)
}
\arguments{
\item{dataset}{The list of object from \code{contingencyTable}}

\item{class_n}{A natural class of land use}

\item{class_m}{A antropic class of land use}
}
\value{
Intensity objects including the levels tables and strationarity test tables
}
\description{
It's a function that perform the intensity test analysis from a table contingency
}
