% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityAnalysis.R
\name{intensityAnalysis}
\alias{intensityAnalysis}
\title{Perform the intensity analysis from a cross-tabulation matrices}
\usage{
intensityAnalysis(dataset, class_n, class_m, area_km2 = TRUE)
}
\arguments{
\item{dataset}{list. The result object from \code{\link{contingenceTable}}.}

\item{class_n}{character. the gaining category in the transition of interest \emph{n}.}

\item{class_m}{character. the losing category in the transition of interest \emph{m}.}

\item{area_km2}{logical. Whether the change would be computed in km2 or quantity of pixel unit.}
}
\value{
Intensity objects including the levels tables and strationarity test tables
}
\description{
This function inplement the Intensity Analysis, a quantitative method to analyze
maps of land categories from several points in time for a single site by
considering cross-tabulation matrices, where one matrix summarizes the change
in each time interval.
}
\details{
There are three levels of analysis, starting from general to more detailed levels, where each level
exposes different types of information given the previous level of analysis
\cite{(Aldwaik and Pontius, 2012, 2013)}.

\enumerate{
 \item The \emph{interval level} examines how the size and speed of change vary
across time intervals.
 \item The \emph{category level} examines how the size and intensity of gross losses
and gross gains in each category vary across categories for each time interval.
 \item The \emph{transition level} examines how the size and intensity of a category’s
transitions vary across the other categories that are available for that transition.
  }


At each level, the method tests for stationarity of patterns across time intervals.

\bold{The function return a list with 6 objects:}
\enumerate{
 \item lulc_table: \code{tibble}. A table of contingency between all step of time analysed, containing 6 columns:
   \enumerate{
   \item Period:  \code{<fct>}. The period of change in the format \code{first year - last year}.
   \item From: \code{<fct>}. The class in the first year.
   \item To: \code{<fct>}. The class in the last year.
   \item km2: \code{<dbl>}. The quantity in kilometer that transit from the classes \code{From}
   to \code{To} in the period.
   \item QtPixel: \code{<int>}. The quantity in number of pixel that transits from
   the classes \code{From} to \code{To} in the period.
   \item Interval: \code{<int>}. The number of time point between the first and the last year of the period.

   }

 \item \emph{lv1_tbl}: An \code{\linkS4class{Interval}} object (containing the \emph{St} and \emph{U} values).
 \item \emph{category_lvlGain}: A \code{\linkS4class{Category}} object (gain with the \emph{Gtj} values).
 \item \emph{category_lvlLoss}: A \code{\linkS4class{Category}} object (loss with the \emph{Lti} values).
 \item \emph{transition_lvlGain_n}: A \code{\linkS4class{Transition}} object (gain in \emph{class n} with \emph{Rtin} and \emph{Wtn} values).
 \item \emph{transition_lvlLoss_m}: A \code{\linkS4class{Transition}} object (loss in \emph{class m} with \emph{Qtmj} and \emph{Vtm} values).

  }
}
\examples{
intensityAnalysis(dataset = SL_2002_2014, class_n = "Ap", class_m = "SG", area_km2 = TRUE)


}
\references{
Aldwaik, S. Z. and Pontius, R. G. (2012) ‘Intensity analysis to unify
measurements of size and stationarity of land changes by interval, category, and
transition’, Landscape and Urban Planning. Elsevier B.V., 106(1), pp. 103–114.
\doi{10.1016/j.landurbplan.2012.02.010}.

Aldwaik, S. Z. and Pontius, R. G. (2013) ‘Map errors that could account for deviations
from a uniform intensity of land change’, International Journal of Geographical
Information Science. Taylor & Francis, 27(9), pp. 1717–1739. \doi{10.1080/13658816.2013.787618}.
}
