% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherplots.R
\name{netgrossplot}
\alias{netgrossplot}
\title{A barplot for the Net Gain/Loss and Gross Changes between land use classes for the period analysed}
\usage{
netgrossplot(dataset, legendtable,
  title = "General Changes (2002 - 2014)", xlab = "Land Use Classes",
  changes = c(GC = "Gross chages", NG = "Net Gain", NL = "Net Loss"),
  color = c(GC = "gray70", NG = "#006400", NL = "#EE2C2C"))
}
\arguments{
\item{dataset}{A table of the multi step transition}

\item{legendtable}{A table contains the legend of the land use classes and his respective color}

\item{title}{The title of the plot}

\item{xlab}{Label for the x axe}

\item{changes}{a character list for the changes labels, the default is c(GC = "Gross chages", NG = "Net Gain", NL = "Net Loss")}

\item{color}{a color list for the three types of chages}
}
\value{
A bar plot
}
\description{
A barplot for the Net Gain/Loss and Gross Changes between land use classes for the period analysed
}
