% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot.IntensityL01}
\alias{plot.IntensityL02}
\alias{plot.IntensityL03}
\alias{plot,IntensityL01,ANY-method}
\alias{plot,plot.IntensityL01,ANY-method}
\alias{plot,IntensityL02,ANY-method}
\alias{plot,plot.IntensityL02,ANY-method}
\alias{plot,IntensityL03,ANY-method}
\alias{plot,plot.IntensityL03,ANY-method}
\title{Plot method for intensity objects}
\usage{
\method{plot}{IntensityL01}(x, y, labels = c(leftlabel =
  bquote("Area changes" ~ "(" ~ Km^2 ~ ")"), rightlabel =
  "Change rate (\%)", title = "Level 01 Mudancas todo o Periodo"),
  labs = c(type = "Changes", ur = "Uniform Rate"), marginplot = c(lh =
  -10, rh = 0), leg_curv = c(x = 1/10, y = 1/10), color_bar = c(faster
  = "#B22222", slower = "#006400", area = "gray40"), ...)

\method{plot}{IntensityL02}(x, y, labels = c(leftlabel =
  bquote("Gain/Loss area (" ~ Km^2 ~ ")"), rightlabel =
  "Intensity Gain/Loss (\%)", title = "Level 02 Ganho/Perda Anual"),
  labs = c(type = "Classes", ur = "Uniform Rate"), marginplot = c(lh =
  0.5, rh = 0.5), leg_curv = c(x = 1/10, y = 1/10), ...)

\method{plot}{IntensityL03}(x, y, labels = c(rightlabel =
  bquote("Intensity Gain/Loss of" ~ .(as.character(dataset[[7]][[1]])) ~
  "(\%)"), leftlabel = bquote("Gain/Loss of " ~
  .(as.character(dataset[[7]][[1]])) ~ "(" ~ Km^2 ~ ")"), title =
  "Level 03 Ganho da classe m/n"), labs = c(type = "Classes", ur =
  "Uniform Rate"), marginplot = c(lh = 0.5, rh = 0.5), leg_curv = c(x =
  1/10, y = 1/10), ...)

\S4method{plot}{IntensityL01,ANY}(x, y, labels = c(leftlabel =
  bquote("Area changes" ~ "(" ~ Km^2 ~ ")"), rightlabel =
  "Change rate (\%)", title = "Level 01 Mudancas todo o Periodo"),
  labs = c(type = "Changes", ur = "Uniform Rate"), marginplot = c(lh =
  -10, rh = 0), leg_curv = c(x = 1/10, y = 1/10), color_bar = c(faster
  = "#B22222", slower = "#006400", area = "gray40"), ...)

\S4method{plot}{IntensityL02,ANY}(x, y, labels = c(leftlabel =
  bquote("Gain/Loss area (" ~ Km^2 ~ ")"), rightlabel =
  "Intensity Gain/Loss (\%)", title = "Level 02 Ganho/Perda Anual"),
  labs = c(type = "Classes", ur = "Uniform Rate"), marginplot = c(lh =
  0.5, rh = 0.5), leg_curv = c(x = 1/10, y = 1/10), ...)

\S4method{plot}{IntensityL03,ANY}(x, y, labels = c(rightlabel =
  bquote("Intensity Gain/Loss of" ~ .(as.character(dataset[[7]][[1]])) ~
  "(\%)"), leftlabel = bquote("Gain/Loss of " ~
  .(as.character(dataset[[7]][[1]])) ~ "(" ~ Km^2 ~ ")"), title =
  "Level 03 Ganho da classe m/n"), labs = c(type = "Classes", ur =
  "Uniform Rate"), marginplot = c(lh = 0.5, rh = 0.5), leg_curv = c(x =
  1/10, y = 1/10), ...)
}
\arguments{
\item{x}{an object from the intensity analysis data}

\item{y}{not used}

\item{labels}{character}

\item{labs}{character}

\item{marginplot}{character}

\item{leg_curv}{character}

\item{color_bar}{character}

\item{\dots}{additional arguments}

\item{leftlabel}{character}

\item{rightlabel}{character}

\item{title}{character}

\item{type}{character}

\item{ur}{character}

\item{lh}{character}

\item{rh}{character}
}
\value{
An intensity graph
}
\description{
Plot method for intensity objects
}
