% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherplots.R
\name{sankeyLand}
\alias{sankeyLand}
\title{A sandkey plot for the land use land cover transition for the period analysed}
\usage{
sankeyLand(dataset, legendtable, iterations = 0)
}
\arguments{
\item{dataset}{Table of one step transition for a onestep sankey plot or a table multi step
transition for a multistep sankey plot.}

\item{legendtable}{A table contains the legend of the land use classes and his respective color}

\item{iterations}{numeric. Number of iterations in the diagramm layout for computation
of the depth (y-position) of each node.}
}
\value{
A land use land cover transition sankey plot for the period analysed
}
\description{
A sandkey plot for the land use land cover transition for the period analysed
}
\seealso{
\code{\link[networkD3]{sankeyNetwork}}
}
