% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherplots.R
\name{sankeyLand}
\alias{sankeyLand}
\title{A sandkey plot for the land use land cover transition for the period analysed}
\usage{
sankeyLand(dataset, legendtable, iterations = 0)
}
\arguments{
\item{dataset}{table of one step transition or multi step transition for a onestep or multistep sankey plot}

\item{legendtable}{A table contains the legend of the land use classes and his respective color}

\item{iterations}{see networkD3 link sankeyNetwork}
}
\value{
A land use land cover transition sankey plot for the period analysed
}
\description{
A sandkey plot for the land use land cover transition for the period analysed
}
