% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherplots.R
\name{sankeyLand}
\alias{sankeyLand}
\title{A sankey diagram}
\usage{
sankeyLand(dataset, legendtable, iterations = 0)
}
\arguments{
\item{dataset}{\code{<tibble>}. A table of the multi step transition (\code{lulc_Mulstistep})
or one step transition (\code{lulc_OneStep}) from \code{\link{contingencyTable}}, for a a coresponding sankey diagram.}

\item{legendtable}{\code{<tibble>}. A table (\code{lulc_Onestep}) contains the legend of the LUC category and his respective color.}

\item{iterations}{numeric. Number of iterations in the diagramm layout for computation
of the depth (y-position) of each node \code{\link[networkD3]{sankeyNetwork}}.}
}
\value{
A sankey diagram
}
\description{
A sankey showing the LUC transition for the period analysed.
}
\seealso{
\code{\link[networkD3]{sankeyNetwork}}
}
