% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SL_2002_2014}
\alias{SL_2002_2014}
\title{Tables of Land Use and Land Cover of the São Lourenço River Basin (2002 - 2014)}
\format{A data list with 5 objects:
\describe{
  \item{lulc_Multistep}{\code{<tibble>} A table of contingency between all step of time analysed, contains 8 columns:
  \enumerate{
  \item Period: \code{<chr>} period between the years analyzed
  \item From: \code{<int>} The pixel value from
  \item To: \code{<int>} The pixel value to
  \item km2: \code{<dbl>} The area of change in squared kilometer
  \item QtPixel: \code{<int>} The change in quatity of pixel
  \item Interval: \code{<int>} The interval in years between the period analyzed
  \item yearFrom: \code{<int>} The year that the change come from
  \item yearTo: \code{<int>} The year that the change go for
  }}
  \item{lulc_Onstep}{\code{<tibble>} A table of contingensy of the first and last year analysed with the same columns
  as \code{lulc_Mulstistep} \code{tibble}}
  \item{tb_legend}{\code{<tibble>} A table of the pixel value, their names and colors, contains 3 columns:
  \enumerate{
  \item classValue: \code{<int>} the pixel value of the classes of land use
  \item className: \code{<fct>} The name or legend associated with a given pixel value
  \item color: \code{<chr>} The color associate with the given pixel value
  }}
  \item{totalArea}{\code{<tibble>} A table of two columns with the total area of study, contains 2 columns:
  \enumerate{
  \item area_km2: \code{<dbl>} The total area of study in km2
  \item QtPixel: \code{<int>} The total area of study in quantity of pixel
  }}
  \item{totalInterval}{\code{<int>} A number represinting the whole interval of time analysed}
  }}
\source{
\url{https://www.embrapa.br/pantanal/bacia-do-alto-paraguai}
}
\usage{
data(SL_2002_2014)
}
\description{
A list containing five objects from \code{\link{contingenceTable}} funtion
with the \code{\link{SaoLourencoBasin}} as input
(\code{SL_2002_2014 <- contingenceTable(input_raster = SaoLourencoBasin, pixelresolution = 30)}).
}
\keyword{datasets}
