% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalfunctions.R
\name{summary_dir}
\alias{summary_dir}
\title{Summary of multiple parameters in a raster directory}
\usage{
summary_dir(path)
}
\arguments{
\item{path}{The path for the Raster*/SpatRaster directory or list of Raster*/SpatRaster to be
analysed. Supports raster package objects and terra package objects.}
}
\value{
Table with the raster parameters in columns
}
\description{
Listing major characteristics of raster inputs. Those characteristics are the
dimensions, the resolution, the extent, the values (min, max) and the
coordinate reference system. This function supports both raster and terra objects
for improved performance.
}
\examples{
\donttest{
url <- "https://zenodo.org/record/3685230/files/SaoLourencoBasin.rda?download=1"
temp <- tempfile()
download.file(url, temp, mode = "wb") # downloading the SaoLourencoBasin dataset
load(temp)
# the summary_dir() function, with the SaoLourencoBasin dataset

summary_dir(raster::unstack(SaoLourencoBasin))
}

}
