% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalfunctions.R
\name{summary_map}
\alias{summary_map}
\title{Quick summary of categorical raster raster}
\usage{
summary_map(path)
}
\arguments{
\item{path}{The path for the raster to be analysed}
}
\value{
A table containing in columns each pixel value and its quantity
}
\description{
Quick summary of categorical raster raster
}
\examples{

summary_map(demo_landscape(2000, res = 1, prob = c(0.05, 0.3, 0.05, 0.4, 0.2))[[1]])
}
